+++ packages/taro-fast-common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 3.1.3
+
+### Patch Changes
+
+- update
+
+## 3.1.2
+
+### Patch Changes
+
+- update
+
 ## 3.1.1
 
 ### Patch Changes