+++ packages/taro-fast-component/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 3.0.9
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - taro-fast-common@3.1.3
+
+## 3.0.8
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - taro-fast-common@3.1.2
+
 ## 3.0.7
 
 ### Patch Changes